+++ benchmarks/libraries/dst/staark-patch.base.js
@@ -100,6 +100,10 @@ var updateAttributes = (element, newAttributes, oldAttributes) => {
             }
           } else {
             if (type === "boolean") {
+              if (!value) {
+                element.removeAttribute(name);
+                continue;
+              }
               value = "true";
             } else if (type !== "string") {
               value = value.toString();
@@ -147,7 +151,7 @@ var prepare = (rootElement, oldAbstractTree) => {
             const oldAbstract = oldChildAbstracts[oldIndex];
             if (oldAbstract.t && newAbstract.t === oldAbstract.t || !oldAbstract.t && !newAbstract.t) {
               matched = true;
-              if (newIndex !== oldIndex) {
+              if (newIndex !== oldIndex + newCount) {
                 element.insertBefore(
                   element.childNodes[oldIndex + newCount],
                   element.childNodes[newIndex]
@@ -199,7 +203,7 @@ var prepare = (rootElement, oldAbstractTree) => {
               );
             }
             const insertAdjacentElement = (element2, elementAbstract2, position) => {
-              if (!elementAbstract2 || elementAbstract2.t) {
+              if (position && (!elementAbstract2 || elementAbstract2.t)) {
                 element2.insertAdjacentElement(
                   position,
                   childElement
@@ -219,9 +223,9 @@ var prepare = (rootElement, oldAbstractTree) => {
               );
             } else if (((_a = oldChildAbstracts == null ? void 0 : oldChildAbstracts.length) != null ? _a : 0) + newCount > newIndex) {
               insertAdjacentElement(
-                element.childNodes[newIndex],
-                oldChildAbstracts[newIndex + newCount],
-                "beforebegin"
+                element.childNodes[newIndex]
+                // (oldChildAbstracts as NodeContent[])[newIndex + newCount],
+                // 'beforebegin',
               );
             } else {
               insertAdjacentElement(
@@ -230,11 +234,10 @@ var prepare = (rootElement, oldAbstractTree) => {
                 "beforeend"
               );
             }
-            newCount++;
           } else {
             childElement = typeof newAbstract === "string" ? newAbstract : newAbstract.c;
             const insertAdjacentText = (element2, elementAbstract2, position) => {
-              if (!elementAbstract2 || elementAbstract2.t) {
+              if (position && (!elementAbstract2 || elementAbstract2.t)) {
                 element2.insertAdjacentText(
                   position,
                   childElement
@@ -254,9 +257,9 @@ var prepare = (rootElement, oldAbstractTree) => {
               );
             } else if (((_b = oldChildAbstracts == null ? void 0 : oldChildAbstracts.length) != null ? _b : 0) + newCount > newIndex) {
               insertAdjacentText(
-                element.childNodes[newIndex],
-                oldChildAbstracts[newIndex + newCount],
-                "beforebegin"
+                element.childNodes[newIndex]
+                // (oldChildAbstracts as NodeContent[])[newIndex + newCount],
+                // 'beforebegin',
               );
             } else {
               insertAdjacentText(
@@ -265,8 +268,8 @@ var prepare = (rootElement, oldAbstractTree) => {
                 "beforeend"
               );
             }
-            newCount++;
           }
+          newCount++;
         }
       }
     }