+++ packages/staark-components/src/index.ts
@@ -1,4 +1,0 @@
-export * from './scripts/inputField.js'
-export * from './scripts/inputFile.js'
-export * from './scripts/inputList.js'
-export * from './scripts/router.js'