+++ packages/staark-isomorphic/dst/staark-isomorphic.js
@@ -191,6 +191,14 @@ var fctory = new Proxy({}, {
   }
 });
 
+// ../staark-common/src/match.ts
+var match = (pattern, lookup) => {
+  if (lookup && pattern in lookup && lookup[pattern]) {
+    return arrayify(lookup[pattern]);
+  }
+  return [];
+};
+
 // ../staark-common/src/memo.ts
 var memo = (render, memory) => ({
   _: marker,
@@ -410,6 +418,7 @@ export {
   conditional,
   factory,
   fctory,
+  match,
   memo,
   nde,
   node,