+++ packages/staark-isomorphic/src/index.ts
@@ -1,6 +1,7 @@
 export * from '@doars/staark-common/src/conditional.js'
 export * from '@doars/staark-common/src/factory.js'
 export * from '@doars/staark-common/src/fctory.js'
+export * from '@doars/staark-common/src/match.js'
 export * from '@doars/staark-common/src/memo.js'
 export * from '@doars/staark-common/src/nde.js'
 export * from '@doars/staark-common/src/node.js'