+++ packages/staark-patch/dst/staark-patch.base.iife.js
@@ -114,6 +114,10 @@
               }
             } else {
               if (type === "boolean") {
+                if (!value) {
+                  element.removeAttribute(name);
+                  continue;
+                }
                 value = "true";
               } else if (type !== "string") {
                 value = value.toString();
@@ -161,7 +165,7 @@
               const oldAbstract = oldChildAbstracts[oldIndex];
               if (oldAbstract.t && newAbstract.t === oldAbstract.t || !oldAbstract.t && !newAbstract.t) {
                 matched = true;
-                if (newIndex !== oldIndex) {
+                if (newIndex !== oldIndex + newCount) {
                   element.insertBefore(
                     element.childNodes[oldIndex + newCount],
                     element.childNodes[newIndex]
@@ -213,7 +217,7 @@
                 );
               }
               const insertAdjacentElement = (element2, elementAbstract2, position) => {
-                if (!elementAbstract2 || elementAbstract2.t) {
+                if (position && (!elementAbstract2 || elementAbstract2.t)) {
                   element2.insertAdjacentElement(
                     position,
                     childElement
@@ -233,9 +237,9 @@
                 );
               } else if (((_a = oldChildAbstracts == null ? void 0 : oldChildAbstracts.length) != null ? _a : 0) + newCount > newIndex) {
                 insertAdjacentElement(
-                  element.childNodes[newIndex],
-                  oldChildAbstracts[newIndex + newCount],
-                  "beforebegin"
+                  element.childNodes[newIndex]
+                  // (oldChildAbstracts as NodeContent[])[newIndex + newCount],
+                  // 'beforebegin',
                 );
               } else {
                 insertAdjacentElement(
@@ -244,11 +248,10 @@
                   "beforeend"
                 );
               }
-              newCount++;
             } else {
               childElement = typeof newAbstract === "string" ? newAbstract : newAbstract.c;
               const insertAdjacentText = (element2, elementAbstract2, position) => {
-                if (!elementAbstract2 || elementAbstract2.t) {
+                if (position && (!elementAbstract2 || elementAbstract2.t)) {
                   element2.insertAdjacentText(
                     position,
                     childElement
@@ -268,9 +271,9 @@
                 );
               } else if (((_b = oldChildAbstracts == null ? void 0 : oldChildAbstracts.length) != null ? _b : 0) + newCount > newIndex) {
                 insertAdjacentText(
-                  element.childNodes[newIndex],
-                  oldChildAbstracts[newIndex + newCount],
-                  "beforebegin"
+                  element.childNodes[newIndex]
+                  // (oldChildAbstracts as NodeContent[])[newIndex + newCount],
+                  // 'beforebegin',
                 );
               } else {
                 insertAdjacentText(
@@ -279,8 +282,8 @@
                   "beforeend"
                 );
               }
-              newCount++;
             }
+            newCount++;
           }
         }
       }