+++ packages/staark-patch/dst/staark-patch.iife.js
@@ -205,6 +205,14 @@
     }
   });
 
+  // ../staark-common/src/match.ts
+  var match = (pattern, lookup) => {
+    if (lookup && pattern in lookup && lookup[pattern]) {
+      return arrayify(lookup[pattern]);
+    }
+    return [];
+  };
+
   // ../staark-common/src/nde.ts
   var nde = (selector, contents) => {
     const [type, attributes] = selectorToTokenizer(selector);
@@ -297,6 +305,10 @@
               }
             } else {
               if (type === "boolean") {
+                if (!value) {
+                  element.removeAttribute(name);
+                  continue;
+                }
                 value = "true";
               } else if (type !== "string") {
                 value = value.toString();
@@ -344,7 +356,7 @@
               const oldAbstract = oldChildAbstracts[oldIndex];
               if (oldAbstract.t && newAbstract.t === oldAbstract.t || !oldAbstract.t && !newAbstract.t) {
                 matched = true;
-                if (newIndex !== oldIndex) {
+                if (newIndex !== oldIndex + newCount) {
                   element.insertBefore(
                     element.childNodes[oldIndex + newCount],
                     element.childNodes[newIndex]
@@ -396,7 +408,7 @@
                 );
               }
               const insertAdjacentElement = (element2, elementAbstract2, position) => {
-                if (!elementAbstract2 || elementAbstract2.t) {
+                if (position && (!elementAbstract2 || elementAbstract2.t)) {
                   element2.insertAdjacentElement(
                     position,
                     childElement
@@ -416,9 +428,9 @@
                 );
               } else if (((_a = oldChildAbstracts == null ? void 0 : oldChildAbstracts.length) != null ? _a : 0) + newCount > newIndex) {
                 insertAdjacentElement(
-                  element.childNodes[newIndex],
-                  oldChildAbstracts[newIndex + newCount],
-                  "beforebegin"
+                  element.childNodes[newIndex]
+                  // (oldChildAbstracts as NodeContent[])[newIndex + newCount],
+                  // 'beforebegin',
                 );
               } else {
                 insertAdjacentElement(
@@ -427,11 +439,10 @@
                   "beforeend"
                 );
               }
-              newCount++;
             } else {
               childElement = typeof newAbstract === "string" ? newAbstract : newAbstract.c;
               const insertAdjacentText = (element2, elementAbstract2, position) => {
-                if (!elementAbstract2 || elementAbstract2.t) {
+                if (position && (!elementAbstract2 || elementAbstract2.t)) {
                   element2.insertAdjacentText(
                     position,
                     childElement
@@ -451,9 +462,9 @@
                 );
               } else if (((_b = oldChildAbstracts == null ? void 0 : oldChildAbstracts.length) != null ? _b : 0) + newCount > newIndex) {
                 insertAdjacentText(
-                  element.childNodes[newIndex],
-                  oldChildAbstracts[newIndex + newCount],
-                  "beforebegin"
+                  element.childNodes[newIndex]
+                  // (oldChildAbstracts as NodeContent[])[newIndex + newCount],
+                  // 'beforebegin',
                 );
               } else {
                 insertAdjacentText(
@@ -462,8 +473,8 @@
                   "beforeend"
                 );
               }
-              newCount++;
             }
+            newCount++;
           }
         }
       }
@@ -503,6 +514,7 @@
     conditional,
     factory,
     fctory,
+    match,
     nde,
     node,
     prepare,