+++ packages/staark/dst/staark.js
@@ -191,6 +191,14 @@ var fctory = new Proxy({}, {
   }
 });
 
+// ../staark-common/src/match.ts
+var match = (pattern, lookup) => {
+  if (lookup && pattern in lookup && lookup[pattern]) {
+    return arrayify(lookup[pattern]);
+  }
+  return [];
+};
+
 // ../staark-common/src/memo.ts
 var memo = (render, memory) => ({
   _: marker,
@@ -363,7 +371,7 @@ var mount = (rootElement, renderView, initialState, oldAbstractTree) => {
               try {
                 value(event);
               } catch (error) {
-                console.warn("listener error", error);
+                console.error("listener error", error);
               }
               listenerCount--;
               updateAbstracts();
@@ -405,7 +413,11 @@ var mount = (rootElement, renderView, initialState, oldAbstractTree) => {
               }
             } else {
               if (type === "boolean") {
-                value = value ? "true" : "false";
+                if (!value) {
+                  element.removeAttribute(name);
+                  continue;
+                }
+                value = "true";
               } else if (type !== "string") {
                 value = value.toString();
               }
@@ -441,9 +453,9 @@ var mount = (rootElement, renderView, initialState, oldAbstractTree) => {
   let oldMemoList = [];
   let newMemoList = [];
   const resolveMemoization = (memoAbstract) => {
-    let match = oldMemoList.find((oldMemo) => oldMemo.r === memoAbstract.r && equalRecursive(oldMemo.m, memoAbstract.m));
-    if (!match) {
-      match = {
+    let match2 = oldMemoList.find((oldMemo) => oldMemo.r === memoAbstract.r && equalRecursive(oldMemo.m, memoAbstract.m));
+    if (!match2) {
+      match2 = {
         c: arrayify(
           memoAbstract.r(
             state,
@@ -454,11 +466,11 @@ var mount = (rootElement, renderView, initialState, oldAbstractTree) => {
         r: memoAbstract.r
       };
     }
-    if (!newMemoList.includes(match)) {
-      newMemoList.push(match);
+    if (!newMemoList.includes(match2)) {
+      newMemoList.push(match2);
     }
     return cloneRecursive(
-      match.c
+      match2.c
     );
   };
   const updateElementTree = (element, newChildAbstracts, oldChildAbstracts, elementAbstract) => {
@@ -486,7 +498,7 @@ var mount = (rootElement, renderView, initialState, oldAbstractTree) => {
             const oldAbstract = oldChildAbstracts[oldIndex];
             if (oldAbstract.t && newAbstract.t === oldAbstract.t || !oldAbstract.t && !newAbstract.t) {
               matched = true;
-              if (newIndex !== oldIndex) {
+              if (newIndex !== oldIndex + newCount) {
                 element.insertBefore(
                   element.childNodes[oldIndex + newCount],
                   element.childNodes[newIndex]
@@ -569,7 +581,6 @@ var mount = (rootElement, renderView, initialState, oldAbstractTree) => {
                 "beforeend"
               );
             }
-            newCount++;
           } else {
             childElement = typeof newAbstract === "string" ? newAbstract : newAbstract.c;
             const insertAdjacentText = (element2, elementAbstract2, position) => {
@@ -604,8 +615,8 @@ var mount = (rootElement, renderView, initialState, oldAbstractTree) => {
                 "beforeend"
               );
             }
-            newCount++;
           }
+          newCount++;
         }
       }
     }
@@ -688,6 +699,7 @@ export {
   conditional,
   factory,
   fctory,
+  match,
   memo,
   mount,
   nde,