+++ packages/staark/src/index.iife.ts
@@ -3,6 +3,7 @@ import { iife } from '../../../.scripts/iife.js'
 import { conditional } from '@doars/staark-common/src/conditional.js'
 import { factory } from '@doars/staark-common/src/factory.js'
 import { fctory } from '@doars/staark-common/src/fctory.js'
+import { match } from '@doars/staark-common/src/match.js'
 import { memo } from '@doars/staark-common/src/memo.js'
 import { nde } from '@doars/staark-common/src/nde.js'
 import { node } from '@doars/staark-common/src/node.js'
@@ -15,6 +16,7 @@ iife([
   conditional,
   factory,
   fctory,
+  match,
   memo,
   mount,
   nde,