+++ packages/staark-patch/CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
+## 1.2.0
+
+- `match` function added.
+- Fix bug when morphing nodes.
+- Fix bug when adding new nodes.
+
 ## 1.1.0
 
-- Conditional function added.
+- `conditional` function added.
 
 ## 1.0.1